120-cell is rendering correctly
The issue before was that Jacob's quaternions were in w-first form, so
my rotations were all goofy.  After correcting that, and correcting the
default quaternions to [0,0,0,1], all is well.

The command takes WAY too long, so I need to look for optimizations.  I did
a little profiling, and there are three GC phases, so I'm creating and
discarding far too many objects.