Optimized JSweet rendering
I'm testing the 120-cell command.  Initially, the time was dominated by
rendering, after the generate().

I added the working attribute to the state, and it gets set to true while
all the other dispatches happen to resolve shapes.  Now the generate() is
easily dominating.

I also did a few minor optimizations in JsAN; the call stack is shallower,
but I'm not sure the result is much faster.

Finally, I changed Polytope4d.WythoffListener to use a string as a map
key.  Again, probably not a great deal of improvement, but I think
toString() is a bit faster than hashValue().

Overall, the cost of the generate() function is now strongly dominated
by the goldenField.plus() and .times() functions, so I suspect we're just
about as good as we can get, here, until we build a faster gcd().