Using Promises for JSweet initialization
I'm using async / await in init, and I inlined the body of
setupIcosahedralSymmetry and loadH4Quaternions inside it.  This ensures
that the quaternions
are ready to parse when the GoldenFieldApplication requires them.

However, I really need to kick off a whole collection of promises to
inject all the shape VEFs, though GFA init does not need to wait for them.
I'll want a Promise.all to capture when shapes are ready for icosa symm.
All these injections should use ResourceLoader, and I can deprecate the
specific injection mechanism for ExportedVEFShapes (used for Unity), or
just adapt it to ResourceLoader.