Fixed defect in loading model using custom tetrahedral symmetry
Built-in tetrahedral symmetry was always handled because of the test for
"." in the name of a tool, in AbstractToolFactor.deserializeTool().
However, for custom tetrahedral tools, the names were not prefixed this
way, so there was nothing in the serialization to distinguish tetrahedral
from octahedral or from icosahedral... the edit is always just "SymmetryTool".

The other option to fix this would be to register the TetrahedralToolFactory
explicitly in FieldApplication.registerToolFactories(), which is only for
deserialization, but that feels like a more impactful change than this one.