Updating JOGL to 2.4.0 to fix the NSWindow exception
Due to my upgraded Catalina OS, what used to be a warning about some
native window APIs on Mac OS X has turned into an exception (as they
promised in the warning).  JOGL has a fix, according to Sven G, so I've
tried to upgrade to 2.4.0, although it is not available from Maven yet.

I found the RC builds, and adjusted the "Ivy" repos to pick them up.
I got a runtime error because they do not contain all the native libs
like the "jogl-all-main.jar", etc. that get published to Maven Central.
I manually listed all the necessary natives (7 each for jogl-all and
gluegen-rt), and now the app runs with "gr desktop:run".

In addition, I controlled the jreHome for the createApp task to OpenJDK 8,
and now the packaged Mac app also runs!  (The last release, 7.0.22, had
started to crash on startup, using JRE 1.8.0_152, with a curious NPE
while loading quaternion VEFs.)

I'll have to install OpenJDK 8 on the build machine.  I'll leave the Windows
build alone, for now.