WIP bringing the build-plane experience to the new client
Lots to do still.  First, all the signs are wrong, and so clicking a grid
box creates a ball on the opposite side.  Second, I'm not trying to build
any struts from the grid center.  Finally, the hover behavior is not
implemented yet.

I also need to figure out a mode switch, so that ball clicks don't select.