Polyhedron.Face normal computation removed
Face normals are no longer computed in the AlgebraicField, except on-demand
for Java2d and DAE exporters.

Computing normals in the AlgebraicField can be expensive and failure-prone,
particularly if you don't have arbitrary precision rational numbers.
Working with automatic struts in Javascript (without any port of BigRational),
I actually saw overflow errors, caught when trying to simplify the resulting rationals.
For this reason, I'm disabling normal computation entirely.
This is fine in Java, because ShapeAndInstances already did its own normal computation
in floating point, and the Javascript rendering does the same.
There is simply no reason to compute the normals here.

I have taken this opportunity to remove the OpenGLExporter, which is
completely irrelevant now.