Setting editMode based on query parameter
Bundles/index.js is now checking the query parameter before deciding which
bundles to include.  Either jsweet or vzomejava gets set as the 'model'
state, and both meet the requirements for the modelcanvas selector.

With no query parameter (or url) you get the "slow" (CheerpJ) implementation.

With editMode=plane, you get the new build plane.

Otherwise, you get click-to-select and the edit menu.

Note that workingPlane mode is still buggy.