Fixed automatic orbit bug
John Hall tripped over this one, though I think I may have seen it before.

We have to store all symmetries, not just the current one, due to sequences like this:

  1. drag an icosahedral olive strut
  2. switch to octahedral symmetry
  3. do "build with this" on the olive strut
  4. drag out a strut
  5. switch back to icosahedral symmetry

The end result is that the command in step 4 records the name of an automatic orbit.
That automatic orbit must be captured in the file.

I have therefore added an "OtherSymmetries" element to the top-level.  For
backward-compatibility, this supplements the existing "SymmetrySystem"
element, rather than replacing it, and does not duplicate that system in the XML.