Fixed working plane bugs, and moved resolution to render
I have moved the shape resolution into the rendering phase, since I never
liked it as "derived state", and it was impossible (or just very
complicated) to make the preview state work correctly that way.

The new "geometry.js" component is extracted from ModelCanvas, and resolves
shapes for each instance on each render.  This is made performant by
memoizing the shapes and shapedInstances tables.  It still needs quite a
bit of refactoring, but it seems to work correctly, rendering the 120-cell.

The preview ball (and strut, usually) are added to mesh.shown in the
selector of ModelCanvas.  This needs to be refactored, but it works well.

All three modes are working nicely.

TODO:

- Disable grid hover while dragging the camera, etc.

- Grid hover behavior for off-grid balls

- Define a cleaner abstraction for shape resolution

- Encapsulate jsweet state more cleanly

- Encapsulate mesh state more cleanly