Corrected the reverse cached strut shape defect
The memoized shapedInstances cache has to store the embedded position and
vectors for the instance.  Since opposite struts ([a,b] and [b,a]) both
have the same ID, we have to consistently use one or the other.

The selected state is NOT stored.  Probably the color should not be, either,
or should be overridden with instance color anyway.