Fixes animation capture.
First, the JQuery code used by the HTML no longer worked.  I have replaced
that with a Vue.js implementation from:

  https://github.com/micheleriva/vue-product-spinner

This component is downloaded from a CDN, so the index.html file is now tiny.

Second, the generated images were huge, and this new page does not set
the image height and width.  I'm now setting a maxSize of 1200 when
generating the images.