WIP Restoring the ability to parse Zomod files
I had to make some Zomod-specific changes to XML2AST, but that's OK, since
there is no Zomic use of that class any more.

ZomicEditorPanel now takes an additional boolean, so it can function for
Zomic or Zomod.

I also had to add ANTLR 2 from the testCompile group to the compile group.