Refactored to normalize parsing between JSweet and CheerpJ
cheerpj.js is the combination of the old jre.js and vzomejava.js.

Note that bundles/index.js defines a "java" bundle, either cheerpj or
jsweet.  The React components depend on the java contract, not on cheerpj
or jsweet specifics.

I also refactored the file loading code in files.js, so now a parser callback
function must be passed.