Entire 120-cell now loads
I fixed the SelectByRadius.  This involved eliminating the use of
SubClassIterator from Manifestations, since the former is just too generic
for Javascript to use.  In particular, the "subclass.isAssignableFrom"
mechanism does not work with JSweet.

I also removed the overloaded constructors from FilteredIterator.

The implementation is not completely correct, since the "set color"
command has no effect, and there may be an ineffective "delete", also.