More improvements to parsing vZome in JSweet code
I can now get past the first legacy CommandEdit, BuildAnchoredSegment, in
the vZomeLogo model.  I had to do a little Java refactoring to achieve this,
in particular adding UndoableEdit.Context.createLegacyCommand().

The parse is now failing on a TetrahedralSymmetry command.  It could be
because I skipped the full implementation of JavaDomElement.getChildNodes().
It could also be because the OrbitSet.Field passed to XmlSymmetryFormat
is faked, and there is no octahedral symmetry available.