Simple tool test is working
I can load a file that contains a point symmetry tool creation and use.

I had to do some defensive coding in TransformationTool.java, where we
were assuming that every manifestation has a non-null color.  This is a
difference between Java and Javascript, because in Java ManifestationImpl
updates its color from the RenderedManifestation on getColor().

There may be a similar defect in copy/paste.