Now can load model with copy/paste
This demonstrates the creation of a Java subclass on the Javascript side,
since I am locally implementing ImportColoredMeshJson.  I could do the
same for other Js* classes, though the classes have to be nested inside
jsweet.js init to have access to vzomePkg.

I also had to unbox the integers in Color.parseWebColor(), to avoid an
invalid overload.