Fixed the debugger sync issue
With an extra initial entry in the dbugger past, undo-all would get the
two histories out of sync, since it is an absolute set-to-zero action
rather than a relative one.

The solution (perhaps not the best one, but the simplest) is to simply
clear the dbugger history after the first reducer call.