WIP NPM module for React components
The module will export just the React components for rendering, initially.

Ultimately I'll include the JSweet-transpiled vZome Java code, too.

The component demo is working, except the lighting is a bit off for some
reason.  I used "npm pack" to build the full package, then installed it
into the app project, and it is working there, too!  (Except for the
lighting, again.)