Some refactoring of the debugger
Also, I fixed a defect in the legacyjava interpret method, which would
prematurely exit the interpreter when a branch within a branch had no
successor edit in the outer branch.

I also enabled sourceMaps for rollup, but I haven't really succeeded much
with using them.  The problem is that the JSweet code is very poorly
formatted by rollup.