Snowpack is working for online and react-library
For online, it is using the monorepo pattern, accessing sources from react-library.
It is sorta doing HMR, but not really... the page reloads, but the source
changes are not picked up.  I think I need some HMR pattern in index.html?

Both dev and build work for react-library, and I shipped and tested the
package successfully.  Both dev and build work in online, as well, though
the build is not bundled.  I will try to bundle the build before shipping it.

I still use CRA / Rollup for the react-library build.