WIP Abstracting interpreter and debugger to support JSON not XML
The first step is to wrap the edit DOM elements in edit objects that
implement the protocol required by the interpreter.

The interpreter is therefore generic, not bound to the legacy Java code.
I have moved it to a new "api" module.

Also there are new "parse" and "getDefaultShaper" async functions.  Both
wait for the vZomeJava.coreState promise.  The idea is that the core
vZome Online component API should not show evidence of the transpiled
Java code.  I will continue to enhance this API, abstracting more of the
legacy behaviors.

I have added a hidden support for JSON output: hold down the alt / option
key while clicking on the download icon.

I'm still experimenting with the JSON serialization!!