Fixing NPE in LinearMapTool
LinearMapTool did not do any bindParameters() validation of the selection,
just relying on countsAreValid().  I've added logic to verify that both bases
(or just the single one) are coincident on a single vertex.

Since legacy edits only require the first two struts of a basis to be coincident,
I didn't add any validation in checkSelection(), since that would break existing
models.

Still TODO: validation of linear independence of the first basis, when there are two.