Simplified external resource loading
Now the JavaScript code completely switches out the default ResourceLoader,
so there is never an attempt to use the ClassLoader.  The code is much
cleaner and simpler on both sides.

Note that I broke the Unity resource loading as a result.  I'll have to
figure out how to use setResourceLoader() from the .NET code, when I
finally return to Unity (VR) work.