WIP Share button now goes to Github vzome-designs repo
The "vzome-designs" repo must be created before any models are shared.
I was unable to perform the right steps for creating the repo automatically,
though some parts are there.  It fails when trying to create the blob,
throwing 409 because the repo is empty.  Docs say I should use "PUT .../contents"
to initialize the repo, but that does not seem to be exposed by EGit.

If the vZome Online links are to work, the repo must be enabled for
Github Pages.

STATUS: the content committed to GitHub is perfect.  All that remains is
to improve the dialog UX in vZome, handling failure cases and launching
the results page in Github.

I updated EGit to the latest release.

I have also started recording the access token in .vZome.prefs.