Fix VEF Import failing when 3D vectors passed to QuaternionProjection
VEF has supported dimensions other than 4D for a while, but it was only used with higher dimensions. It now supports 3D correctly, but 2D or 1D require a change to inflateTo4d() which  I have not done at this time, although the potential issue with 2D is noted in the newly added comments.