WIP Switching to BigRationalImpl
I have added BigRationalFactory, and AAF is now decoupled from BRI.
I left a few of the legacy AFs with a default factory, since I already
have versions of them in Javascript.

All tests are passing, though I had to comment
out a couple of parts of FibonacciTest.java, due to the complexities of
the generics.