Fix: golden VEF import and golden vectors in N-gon fields
TODO: Add a regression file that uses golden VEF import
Be sure it subsequently manipulates some of the parts so it will fail if this breaks, not just render wrong which won't be tested in regression.

Should also be able to parse a Polygon5 VEF in higher 5N-gon fields. Test this too.