POV-Ray exporter now uses real numbers
There was no real value to using the algebraic expressions for ANs, and
this means we can get rid of AF.defineMultiplier(), which I have done.

I had to do quite a bit of debugging to get panels to work, and I finally
switched from "union { polygon { ..." to "mesh { triangle { ...".  This
way there is no problem with planarity of polygon vertices.