Create a ProjectionTool from a ball and strut
A ProjectionTool can now be created from a strut that's normal to the projection plane and a ball on the plane.
The original way of creating a ProjectionTool from a panel and an optional strut still works the same as before.
A regression file is included which uses this new capability.