Interpreting vZome XML using setTimeout
This means we will let the browser render in between interpreting edits.
Of course, any give edit could block rendering, so we'll still have some
jank.

Currently, when I profile a heavy document (high-order field, John K's
2021/07/16/00-02-19-60-gon-sqrt-grids/60-gon-sqrt-grids.vZome), I see the
parse taking 10 seconds.  The file has dozens of automatic orbits, and
they all get initialized during createDocument!

The interpret phase is working nicely, using setTimeout to break it up
into tasks.

Rendering (shapeInstances) is still a problem.  It seems to be done with
microtasks, so it is blocking the frame for several seconds.