Optimizing core initialization
I'm trying to defer some part of orbit and zone initialization, making it
lazy, to reduce the initial time.  The laziness is implemented by
Symmetry.getAxis(AV) and symmetry.computeOrbitDots().  The latter is
actually now recomputed each time we do OrbitSetController.recalculateDots().

Ideally, we could make it not just lazy, but actually initialized in a
background thread, so the data is ready when needed to handle an event
for the UI.  In fact, we can still do that, outside of core, by calling
those methods that actually force the init.  (We don't want multi-threaded
code in core init, since it won't transpile with JSweet.)

The main optimization is to defer the matrix transforms in Direction
(Orbit) construction.

I'm leaving createOtherOrbits() in place, since it won't have a high cost
now.  Importantly, the withCorrection() computation has been deferred also.

This optimization reduced the cost of parse() in one of JohnK's big 60-gon
models from 10s to 800ms.

https://raw.githubusercontent.com/John-Kostick/vzome-sharing/main/2021/07/16/00-02-19-60-gon-sqrt-grids/60-gon-sqrt-grids.vZome