Adding PWA support
I have used Workbox to implement a basic service worker with caching.
I can install the app from either Chrome or Brave (from production),
and it even works offline, except it cannot find the model files.

In Safari on iPhone, it seems to ignore the icons from the manifest, but
perhaps it needs a size I haven't provided.

Note that only the production build has been modified to build the service
worker bundle, which cannot be in ESM format.