WIP Using web worker to run legacy code
I can package a worker, and get messages back and forth.

I can successfully bundle the worker with esbuild, including the legacy
code required to parse.  However, the call to DOMParser fails, of course.
I found some options, googling for 'domparser in web worker'.

I also tried parsing on the main thread with DOMParser, but then the
DOMDocument won't clone to send to the web worker.  NP, it is better to
parse in the worker anyway!

What I really want to do in the worker is *all* editing, legacy or not,
so I'll need a different architecture.  Probably I want the whole Redux
store to be inside the worker, and just send back render changes like I
did for client-server (with Unity).

I have implemented part of this, for previewing .shapes.json.  The preview
model is fetched and parsed in the worker, and the resulting scene is
posted back piecemeal, first the lighting, camera, and embedding in one
step, then each shape, then each instance.