Worker is now launched as early as possible
The web component is now using a dynamic import for loading the React
component, so React and three.js code (3MB un-minimized) is deferred until
the worker has been kicked off.

This means that the .vZome and .shapes.json files are now fetched on the
worker while the browser is also fetching the code to render.