Created design concept to decouple from custom element and React component
The design concept corresponds to the DocumentModel, or perhaps the
DocumentController, from the Java implementation.  The Worker itself is
an implementation detail.

The current test page successfully loads two models (four files) in
parallel while the large dynamic import is still fetching the Javascript
for React, three.js, etc. needed to render them.