WIP Experimenting with using RealizedModelImpl and RenderedModel
The context here is my ongoing work to move the online code to an event-
based rendering model, since that is essentially required by the idea of
running commands in a worker.

Since the Java code already works this way, I'm experimenting with the
idea of removing the "mesh" state managed by Adapter and JsRealizedModel,
and replacing it with the actual legacy RealizedModel and RenderedModel.
Another advantage of this approach would be to (hopefully) eliminate all
those subtle defects with JsRealizedModel and JsSelection.

This represents a fork in the road: should I move to Strings as keys in
RealizedModel?  Will that sacrifice the backward compatibility?  Or should
I abandon this approach and generate render events from the mesh state changes?
I'm trying the first fork, as I've already switched to string keys.

At this time, shapes and instances are flowing back apparently correctly.
I am doing a batch rendering after the XML is interpreted.
However, there are some problems:

1. the camera is not right, missing the "fov" attribute, so nothing
is visible.  I've hacked this for the moment

2. One green strut in vZomeLogo is not getting removed