Worker interpretAndRender is working now for vZomeLogo etc
The extra green strut in vZomeLogo revealed a couple of issues, mostly
having to do with unique keys (signatures) for Manifestations in
RealizedModelImpl, in both the main collection and in the per-edit
collection.

I've added Construction.getSignature() to get a canonical key string for
a point, segment, or polygon.  This then becomes the key for the RMI
collections.

BEWARE: in JSweet (j4ts actually), it seems that Arrays.sort() is NOT
trustworthy on string arrays.  Polygon.getSignature() may not be good!