Removed sceneListener callback from batch file loads
Since the scene is sent as just two pieces, first the camera etc. then
later the geometry, there is no need to use a callback approach for the
sceneListener.  It is now called twice, for both the preview load and the
legacy XML load.

I also removed all the async/await from API interpret(), since that was
really just for using microtasks in the browser context.