New viewer mode for app is working
If the "?url=..." is provided, the app now defaults to a single-design
viewer, equivalent to the web component but with additional capabilities
from the app bar.  Everything in the app bar is currently disabled, but
eventually this mode will support a more explicit "switch to editor"
mechanism, possibly within the viewer component itself.

For the moment, this is a good first step toward using the new worker-based
approach in vZome Online.  Longer term, this is my intended way to offer
support for designs shared in this way before vZome could generate custom
pages; I don't want vZome Online to be a "viewer / editor", just an editor.
I'll use a new query param for the "edit in vZome Online" mechanism.

Lots of stuff is currently disabled, for now, but the "editor/debugger"
opens OK but without any model.

I'm ditching Redux in the browser context, switching to React hooks for
the little bit of state that will be outside the worker.  In particular,
I'll have a list of designs as app state.

I'm also moving away from default exports, on Lucas' advice.