WIP Getting openUrl to work
I have decided to dispense with the idea of "multiple open designs", since
it confuses users, and just having one design under edit is much simpler
to understand and implement.  There is now just one worker, managed by
the "controller".

I also got rid of the share button in the appbar, now redundant with the
download button in the viewer.

The web component is working, and the app is working both in "legacy
viewer" mode and in editor mode for the predefined models.  Remote URL
works OK when there is no preview available.  Local files are not working yet.