Spinner is working
I had some trouble with an event cycle for a while; see comments in folder.jsx.
Due to that, the app comes up empty, with a blank canvas rather than an
initial design.  That's OK, since it needs to open on a new design anyway.

Also, an animated 3D scene would make a better spinner.