WIP Restoring undo/redo and alerts
Alerts are working, just using useState at the app level.

The undo/redo buttons have been moved inside DesignEditor, out of the
app bar.  The debugger is also moved inside the DesignEditor.

The view and controller are hooked up to the worker, but the worker is
not taking any actions, and it always sets the canUndo/Redo properties
to false.