Adopted redux-in-worker to normalize state
Redux-in-worker seems to be exactly what I need, providing a highly uniform
model of "remote" state in the worker.  It is simplifying the UI code
considerably, in combination with react-redux hooks.  Note that only diffs
are sent across as messages.

This is very promising for adapting the legacy controller framework,
perhaps in combination with Redux combineReducers().

Answered Questions:

- What happens to the return value from middleware?  Returned from the
   dispatch() call, which means almost always ignored.  It has nothing
   to do with the reducers.
- how does Promise.then().catch().then() work?  As you would expect; it
   is NOT a way to branch the task tree... more like a "finally".
- When should I just start work in then(), as opposed to reporting the
   value to run through middleware again?  Almost always; the promise
   chain will be much easier to read and debug.

Status:

The app works in preview and normal modes.  Undo/redo is still not hooked
up, or the debugger.  The web component is working now also.
The BHall and Article pages seem to work, up to CORS errors downloading
their models.

I have a disparity between the dev build and the prod build, w.r.t.
the "/app" element in the path.  I'm not sure what to do about it.  The
code for loading models must be robust in this case!  (Module loading is
fine, since all modules are at /modules/*.)