WIP Introduced DesignHistoryInspector
This actually replaces DesignEditor, since we won't need that until we
have actual editing supported.

For legacy "?url=..." support, the inspector opens dedicated to the given
design.  If no URL is passed, the file load icon shows.

In all cases, the history panel will show!

STATUS:

I tried redux-in-worker, but I continued to have null instances in the
snapshots when trying to render, and other rendering defects, and I
got fed up with trying to guess how redux-in-worker is computing diffs.

Now I've gone back to just React on the
UI side to hold the state to render, with change events sent from the
worker as I did in the client-server and VR work.

I suspect the right approach will even benefit desktop vZome, for example
by supporting a history debugger in a browser window, with a websocket
replacing the worker.  Perhaps this could even provide a transition path
for desktop-to-web.

Everything is working fine, except for a few problems:

1. branch edits don't get undone

2. debugger controls are not hooked up (and should probably be removed)

3. Tree focus and select are different actions.  Looks like I'll need to
   upgrade MUI if I want to have an onFocusNode() action.

4. On initial load, there is no focused or selected tree item.