WIP Regression test of John Kostick files
The component test page now fetches all (or many) vZome designs from
John's vzome-sharing repo, then sets each URL in turn as the src attribute
of the vzome-viewer.  A 3-second sleep is commented out at the moment.
The failure cases are not handled well, and I also want to add visual
output on the page for successes and failures.

I have also changed my approach for the app when a URL query param is
provided: it won't show the history inspector.  This is for a better
experience for any old shared URLs that may still be used, since it turns
out the tree view is pretty slow to render.

I think the history inspector is broken at the moment.