Restored HistoryInspector function
I also moved the Spinner into the DesignViewer, where it is controlled by
a `useSpinner` property.  This means that the whole viewport is no longer
blocked by the spinner backdrop.

The worker is now once again rendering snapshots for all edits.  It does
not seem to materially impact the performance.