Render snapshots are now before and after edits
The tree remains unchanged, but the semantics of selecting a treenode can
be adjusted by working with either the `before` snapshots or the `after`
snapshots.  This very much improves the semantics of the RenderHistory,
acknowledging that history is "between" the edits.

The HistoryInspector now shows a "--START--" node, rather than the
opposite, and selecting a node shows the state after that edit is applied.