Using a shadow DOM slot to render fallback content
The "light DOM" generated for a `vzome-viewer` element is an `<img>` tag,
originally intended to show when Javascript is disabled, or any other
reason that the web component might not work.  By rendering a `<slot/>`,
we can also choose to show that content if anything goes wrong with
fetching, interpreting, and loading the 3D scene.

For example, this gives a better experience in Firefox, where the lack of
support for module workers means nothing will be rendered.