Add "printable" style
This avoids breaking older models using ReplaceWithShape, since the style
names continue to represent stable shapes.  It is not a perfect solution,
but it is acceptable.  We need to remember to add styles rather than
changing shapes.

I also fixed some ordering issues with the style menus, due to the
placement of AbstractSymmetryPerspective.setDefaultGeometry() calls.