Refactored AWT dependencies out of Controller and DefaultController
I added GraphicsController and DefaultGraphicsController, which extend
the base interface and class with things that depend on AWT.

Everything (non-Swing) that depends on GraphicsController or AWT itself
is now moved to com.vzome.desktop.awt.  All "clean" controllers are moved
to com.vzome.desktop.controller.

Controller itself is moved to com.vzome.desktop.api.

The old "impl" package that held all the controllers is now gone.

The changes are quite extensive, but really all mechanical, implemented
using Eclipse refactoring.  The only oddity is that SymmetryController
is not itself directly using GraphicsController, but it does contain
OrbitSetControllers that do, so it is in the awt package with the others.