Made four separate app entry points to simplify routing
The react-router mechanism was not really a good way to achieve the
separation of the online, browser, fivecell, and bhall/basic apps.  It
was difficult or impossible to set it up to work well both in the dev
server and in production.

For a better solution, I have simply created separate index.html entry
points for the four apps, and separate corresponding index.jsx entry
points.  Thanks to esbuild, I still get a single build step, and maximum
reuse of the Javascript bundles.

The special rule in .htaccess is not necessary for now, since I've removed
all use of react-router.  I'm leaving the .htaccess file though, since
I'm sure I'll need react-router eventually.  With separate entry points,
it should be easy to add separate rules to handle the routing.