Supporting article/lesson viewing in vzome-viewer
I've added SceneMenu to the DesignViewer component, as well as a
toggle to disable it.  It won't work for JSON previews, of course, so we
always fetch the actual XML now.  If the toggle is on, we replace
the preview scene seamlessly.

I have completely decoupled SCENE_RENDERED, so it stays uniform for all
use-cases.  The full, indexed DOM tree and possible article snapshots
are returned separately in DESIGN_INTERPRETED.