Fixed the preview-then-interpret defect
In order to support showing snapshots with a preview JSON, the online
code was rendering twice, once for the preview and again after interpreting
the vZome XML.  In general we want to avoid that, since the interpretation
may have bugs.

Now parseAndInterpret() takes a new "render" argument that defaults to
true.  If a preview JSON was found and rendered, then we call
parseAndInterpret with render=false.

I also added a toggle to disable showing the snapshots menu.  This was
necessary for one of the bhall/basic models.  In the process, I added a
"config" property to the DesignViewer component and its wrappers,
subsuming the old "useSpinner" property.