Correcting the synchronization defect in article mode
This fix is working well, but I don't fully understand why, yet.  What is
clear is that the prior code was setting up a parallel background thread
without synchronization, which clearly caused trouble.  Also, the code
in PagelistPanel was scheduling EDT work from an EDT thread... again,
with no clear synchronization.

I've added extensive comments.  I'm committing this not because I have a
deep understanding, but simply because the results are better than they
were, and I need to set this aside right now.  There is more work to do
here, and in general, to take better control of concurrency with
proper synchronization.

Incidentally, I have also removed the old SwingWorker, since Swing now
includes a natively-implemented one.  This does seem to add some snap to
startup and opening a new window!