Disabled preview when showSnapshots is set
I have added a "show-scenes" attribute to the web component, but it is
not working correctly yet.

I have disabled snapshots by default everywhere, since it causes regressions.
It will be opt-in henceforward, using "show-scenes" on the web component.

I've tried to rationalize the contracts for the URL_PROVIDED event to the
store, for the React components, and for the web component.  Requesting
the scenes menu will preclude rendering the preview, so that the worker
code remains simple as before... either preview or XML, never a mix.