Fixed "extra thumbnail" defect in article mode
The problem was one of poorly understood concurrency, and a somewhat
disjointed approach to MVC.

The LessonPanel is now created for every DocumentFrame, and it does not
attempt to update its page state on creation.  This simplifies the model
so that "has.pages" is not so critical, and we have less concern over
the model state when creating the UI.  In general, I think this is a good
approach that will allow concurrency between the EDT creating the UI and
a worker thread loading the document.

Loading the XML for the article now fires property change events, exactly
as if the user were capturing pages, secure in the knowledge that the
listeners will handle these events in order on the EDT, since the
LessonController is using SwingPropertyChangeSupport and the PCEs are
fired from a worker thread.  (SwingPropertyChangeSupport will notify
listeners synchronously if it is running on the EDT already, but from a
worker thread it will always use SwingUtilities.invokeLater().)

Thumbnails are still not rendering, but I doubt that is a threading issue.