Fixed premature lesson update
Document load must complete before the LessonModel fires any properties
to update the UI, so I added a new firePropertyChanges() method.

The UI is very flickery, since it renders each page in the main view as
the PCEs are handled.  I need to tune it so that it simply presents the
first page.