Fixed the selection behavior in the HistoryInspector
I'm working on restoring debugger functionality.  As a start, the
history tree must be able to show where the interpretation failed,
which means the MUI TreeView has to be "controlled".  The selected
node is now completely controlled by the `edit` field in the Redux store,
which is further controlled by the results of interpretation or
"EDIT_SELECTED" actions.

The worker contract is still a bit sloppy around the concept of "before"
or "after" edits.  The worker will respond with scene for either, but
the problem is that URL_PROVIDED and FILE_PROVIDED don't imply one or
the other.

The interpretation is correctly reporting the last successful edit now,
but I still don't have a way to re-trigger the next edit to debug it,
as I did before I switched to the worker.