Restored debug=true capability
The history inspector UI is fine as it is, without player controls;
"go to breakpoint" semantics are more than sufficient.  I may add classic
debug controls with the keyboard at some point, but I don't see any
benefit of mouse controls over just clicking on the tree.  (Persistent
breakpoints are useful in languages with loops and functions, but not
here, where interpretation is so linear.)

In the worker, in the interpret() function, I'm capturing the ongoing
value of "edit" in the RenderHistory, as a high watermark.  This is so
I have a starting point for interpret() when I don't find a scene for
a given id.  Now getScene() is doing a to-breakpoint interpret when a
scene is not already present.  The breakpoint is always transitory.

I had added a new `/debugger` app, but I've concluded it was just a
complication, since the UI is unchanged.