Fix tetrahedral symm defect in rootTwo for Online
This was a beast.

The defect was in SymmetryTool, where the switch statement wasn't handling
"synestructics" as a symmetry name.  That switch statement itself, of
course, is an egregious hack.  There should be subclasses for
OctahedralSymmetryTool and TetrahedralSymmetryTool.  Due to this missing
switch case, the tetrahedral symmetry case was falling through to the
default, producing 23 transformations rather than 11.

Furthermore, the defect was obscured on desktop vZome because the
ToolsModel map resolved a collision differently.  The ToolsModel pools
all tools from all SymmetrySystems, making the assumption that
"tetrahedral.whatever" means the same thing in all systems.
Due to the defect above, this was not true.  In the online code, the
Javascript implementation of ToolsModel resolved the conflict differently.