Added side-effects exporter
In order to support full, correct interpretation of the "history" export,
I needed to include the symmetry systems and tools in the export.
However, that would break the regression tests, introducing a difference,
so I've created a new SideEffectsExporter for the purpose.

Note that the file extension ends in ".vZome", so I don't need a special
case in the online debugger file pattern.